changing filenames to ssssnake case
