+++ .gitignore
@@ -1,2 +1,3 @@
 /**/.DS_Store
 /coverage
+/lib/twilio_auth.env
