+++ Gemfile
@@ -3,6 +3,7 @@ ruby '2.3.1'
 
 gem 'capybara'
 gem 'coveralls', require: false
+gem 'dotenv'
 gem 'rake'
-
 gem 'rspec'
+gem 'twilio-ruby'
