+++ lib/phone.rb
@@ -1,0 +1,6 @@
+require 'dotenv'
+class Phone
+  def initialize
+    Dotenv.load
+  end
+end
