+++ spec/phone_spec.rb
@@ -1,0 +1,5 @@
+require "phone"
+describe Phone do
+  describe "#creation "
+  describe "#"
+end
